backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# LLM / embeddings
litellm==1.48.0
openai==1.50.0

# Vector DB
qdrant-client==1.11.3

# Graph
networkx==3.3
python-louvain==0.16

# NLP / ML
transformers==4.44.2
sentence-transformers==3.1.1
torch==2.4.1+cpu

# PDF / document parsing
PyMuPDF==1.24.10
python-docx==1.1.2
trafilatura==1.12.0
grobid-client-python==0.1.4
lxml==5.3.0
beautifulsoup4==4.12.3

# Database
sqlalchemy==2.0.35

# Task queue
celery==5.4.0
redis==5.1.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
numpy==1.26.4
